refactor: move routes and layouts to dedicated folder
