+++ apps/web/src/layouts/Layout.tsx
@@ -1,4 +1,5 @@
 import { Outlet } from "react-router";
+import LogoutButton from "../components/LogoutButton";
 
 export default function Layout() {
   return (
@@ -14,6 +15,15 @@ export default function Layout() {
             </li>
             <li>
               <a href="/register">Register</a>
+            </li>
+            <li>
+              <a href="/app">App</a>
+            </li>
+            <li>
+              <a href="/secret">Secret</a>
+            </li>
+            <li>
+              <LogoutButton />
             </li>
           </ul>
         </nav>
